requirements.txt
# Core ADK and Google Cloud dependencies
google-adk>=1.9.0
a2a-sdk==0.2.16
httpx>=0.24.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.10.0
vertexai>=1.38.0

# Optional integrations for production
firebase-admin>=6.2.0
pymongo>=4.5.0
openai>=1.3.0
newrelic>=8.10.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
